# backend/resources/openapi.yml
openapi: "3.0.1"
info:
  title: Things2Do
  version: 1.0.0
  description: Distributed todo application for recording tasks to be completed.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
paths:
  /groups:
    get:
      tags:
        - Group
      summary: "Fetch all available groups"
      description: "Fetchs a list of all available groups"
      responses:
        "200":
          description: "Sucessfully fetched all available groups"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
              example:
                - id: "507f1f77bcf86cd799439011"
                  name: "Gardening"
                  description:
                    "Contains all the tasks that arise in the garden\n
                    and around the house."
    post:
      tags:
        - Group
      summary: "Creates a new group"
      description: "Creates a new group for grouping tasks"
      requestBody:
        $ref: "#/components/requestBodies/GroupCreatePayload"
      responses:
        "201":
          description: "Successfully created the new group"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
              example:
                id: "507f1f77bcf86cd799439011"
                name: "Gardening"
                description: "Contains all the tasks that arise in the garden\n
                  and around the house."
        "400":
          $ref: "#/components/responses/BadRequestError"
  /groups/{groupId}:
    get:
      tags:
        - Group
      summary: "Fetch a specific group"
      description: "Fetch information about a specific group"
      parameters:
        - name: "groupId"
          in: "path"
          description: "Unique identifier of the group"
          required: true
          schema:
            type: "string"
            format: "id"
            example: "507f1f77bcf86cd799439011"
      responses:
        "200":
          description: "Sucessfully fetched the requested group"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
              example:
                id: "507f1f77bcf86cd799439011"
                name: "Gardening"
                description: "Contains all the tasks that arise in the garden\n
                  and around the house."
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
    patch:
      tags:
        - Group
      summary: "Update a specific group"
      description: "Update fields of a specific group"
      parameters:
        - name: "groupId"
          in: "path"
          description: "Unique identifier of the group"
          required: true
          schema:
            type: "string"
            format: "id"
            example: "507f1f77bcf86cd799439011"
      requestBody:
        $ref: "#/components/requestBodies/GroupUpdatePayload"
      responses:
        "200":
          description: "Sucessfully updated the requested group"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
              example:
                id: "507f1f77bcf86cd799439011"
                name: "Gardening and design"
                description: "Contains all the tasks that arise in the garden\n
                  and around the house."
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
    delete:
      tags:
        - Group
      summary: "Delete a specific group"
      description: "Delete a specific group"
      parameters:
        - name: "groupId"
          in: "path"
          description: "Unique identifier of the group"
          required: true
          schema:
            type: "string"
            format: "id"
            example: "507f1f77bcf86cd799439011"
      responses:
        "204":
          description: "Sucessfully deleted the requested group"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
components:
  schemas:
    Error:
      type: "object"
      properties:
        code:
          type: "string"
        message:
          type: "string"
      required:
        - code
        - message
    Group:
      type: "object"
      properties:
        id:
          type: "string"
          format: "id"
        name:
          type: "string"
          maxLength: 200
        description:
          type: "string"
          maxLength: 2000
      required:
        - id
        - name
...
  requestBodies:
    GroupCreatePayload:
      description: "Fields of new group to add"
      required: true
      content:
        application/json:
          schema:
            type: "object"
            properties:
              name:
                type: "string"
                maxLength: 200
              description:
                type: "string"
                maxLength: 2000
            required:
              - name
          example:
            name: "Gardening"
            description: "Contains all the tasks that arise in the garden\n
              and around the house."
    GroupUpdatePayload:
      description: "Fields of new group to update"
      required: true
      content:
        application/json:
          schema:
            type: "object"
            properties:
              name:
                type: "string"
                maxLength: 200
              description:
                type: "string"
                maxLength: 2000
          example:
            name: "Gardening and design"
...
  responses:
...
    BadRequestError:
      description: "Used when the client provides incorrect information in the\n
        request."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFoundError:
      description: "Used when a resource could not be found."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"